[FIX] file structure
