+++ src/main/java/com/server/Dotori/model/member/controller/MemberController.java
@@ -2,7 +2,7 @@ package com.server.Dotori.model.member.controller;
 
 import com.server.Dotori.model.member.dto.MemberDto;
 import com.server.Dotori.model.member.dto.MemberLoginDto;
-import com.server.Dotori.model.member.service.EmailService;
+import com.server.Dotori.model.member.service.email.EmailService;
 import com.server.Dotori.model.member.service.MemberService;
 import com.server.Dotori.response.ResponseService;
 import com.server.Dotori.response.result.CommonResult;
