+++ src/main/java/com/server/Dotori/model/member/repository/MemberRepositoryCustom.java
@@ -1,0 +1,12 @@
+package com.server.Dotori.model.member.repository;
+
+import com.server.Dotori.model.member.dto.selfstudy.SelfStudyStudentsDto;
+
+import java.util.List;
+
+public interface MemberRepositoryCustom {
+
+    List<SelfStudyStudentsDto> findBySelfStudyAPLLIED();
+
+    void updateSelfStudyStatus();
+}
