--- src/main/java/com/server/Dotori/model/member/service/EmailSendService.java
@@ -1,8 +1,0 @@
-package com.server.Dotori.model.member.service;
-
-import com.server.Dotori.model.member.dto.EmailDto;
-
-public interface EmailSendService {
-    void sendEmail(String userEmail,String key);
-    void sendPasswordEmail(String userEmail, String password);
-}
