--- src/main/java/com/server/Dotori/model/member/service/EmailService.java
@@ -1,11 +1,0 @@
-package com.server.Dotori.model.member.service;
-
-import com.server.Dotori.model.member.Member;
-import com.server.Dotori.model.member.dto.EmailDto;
-import com.server.Dotori.model.member.dto.MemberEmailKeyDto;
-
-public interface EmailService {
-    String authKey(EmailDto emailDto);
-    String authCheck(MemberEmailKeyDto memberEmailKeyDto);
-    Member authPassword(EmailDto emailDto);
-}
