+++ src/main/java/com/server/Dotori/model/member/service/email/EmailService.java
@@ -1,0 +1,11 @@
+package com.server.Dotori.model.member.service.email;
+
+import com.server.Dotori.model.member.Member;
+import com.server.Dotori.model.member.dto.EmailDto;
+import com.server.Dotori.model.member.dto.MemberEmailKeyDto;
+
+public interface EmailService {
+    String authKey(EmailDto emailDto);
+    String authCheck(MemberEmailKeyDto memberEmailKeyDto);
+    Member authPassword(EmailDto emailDto);
+}
