+++ src/main/java/com/server/Dotori/model/member/service/point/PointService.java
@@ -1,0 +1,8 @@
+package com.server.Dotori.model.member.service.point;
+
+import com.server.Dotori.model.member.dto.point.PointDto;
+
+public interface PointService {
+
+    void point(PointDto pointDto);
+}
