+++ src/main/java/com/server/Dotori/model/member/service/selfstudy/SelfStudyService.java
@@ -1,0 +1,16 @@
+package com.server.Dotori.model.member.service.selfstudy;
+
+import com.server.Dotori.model.member.dto.selfstudy.SelfStudyStudentsDto;
+
+import java.util.List;
+
+public interface SelfStudyService {
+
+    void requestSelfStudy();
+
+    void cancelSelfStudy();
+
+    List<SelfStudyStudentsDto> getSelfStudyStudents();
+
+    void updateSelfStudyStatus();
+}
