--- src/main/java/com/server/Dotori/model/point/enumType/Kind.java
@@ -1,5 +1,0 @@
-package com.server.Dotori.model.point.enumType;
-
-public enum Kind {
-    PLUS_POINT, MINUS_POINT
-}
