--- src/main/java/com/server/Dotori/model/point/repository/PointRepository.java
@@ -1,9 +1,0 @@
-package com.server.Dotori.model.point.repository;
-
-import com.server.Dotori.model.point.Point;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PointRepository extends JpaRepository<Point, Long> {
-}
