--- src/main/java/com/server/Dotori/model/point/service/PointService.java
@@ -1,8 +1,0 @@
-package com.server.Dotori.model.point.service;
-
-import com.server.Dotori.model.point.dto.PointDto;
-
-public interface PointService {
-
-    void point(PointDto pointDto);
-}
