--- src/main/java/com/server/Dotori/model/selfstudy/dto/SelfStudyStudentsDto.java
@@ -1,18 +1,0 @@
-package com.server.Dotori.model.selfstudy.dto;
-
-import com.server.Dotori.model.member.enumType.SelfStudy;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class SelfStudyStudentsDto {
-
-    private String stdNum;
-    private String username;
-
-    public SelfStudyStudentsDto(String stdNum, String username) {
-        this.stdNum = stdNum;
-        this.username = username;
-    }
-}
