--- src/main/java/com/server/Dotori/model/selfstudy/repository/SelfStudyRepository.java
@@ -1,9 +1,0 @@
-package com.server.Dotori.model.selfstudy.repository;
-
-import com.server.Dotori.model.selfstudy.SelfStudy;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SelfStudyRepository extends JpaRepository<SelfStudy, Long>, SelfStudyRepositoryCustom {
-}
