--- src/test/java/com/server/Dotori/model/point/PointTest.java
@@ -1,2 +1,0 @@
-package com.server.Dotori.model.point;public class PointTest {
-}
